+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-05-05 13:09:31
-// Commit: docs(storage): update readme and API contracts for storage
+// Updated at: 2026-05-06 12:46:08
+// Commit: feat(storage): integrate custom telemetry logger in storage
 
 class storageComponentState {
-  final String commitRef = '881';
-  final String lastMessage = 'docs(storage): update readme and API contracts for storage';
-  final DateTime timestamp = DateTime.parse('2026-05-05T13:09:31+05:00');
+  final String commitRef = '933';
+  final String lastMessage = 'feat(storage): integrate custom telemetry logger in storage';
+  final DateTime timestamp = DateTime.parse('2026-05-06T12:46:08+05:00');
 }